+++ 9-oh-my-rpg/rsrc-backgrounds/src/index_spec.ts
@@ -12,7 +12,10 @@ describe(`${LIB} - rsrcs`, function() {
 				expect(artwork.source.length, 'source').to.be.above(12)
 			})
 			it('should be manually aligned', () => {
-				expect(artwork.position_pct.x !== 50 || artwork.position_pct.y !== 50, 'alignment').to.be.true
+				if (artwork.position_pct.x === 50 && artwork.position_pct.y === 50) {
+					expect(artwork.position_pct.x, 'alignment x').to.not.equal(50)
+					expect(artwork.position_pct.y, 'alignment y').to.not.equal(50)
+				}
 			})
 			it('should have a lowercase CSS class', () => {
 				expect(artwork.css_class.toLowerCase()).to.equal(artwork.css_class)
@@ -21,7 +24,19 @@ describe(`${LIB} - rsrcs`, function() {
 	})
 
 	describe('copy/past error prevention', function() {
-		it('shouldn’t have duplicated url')
-		it('shouldn’t have duplicated css classes')
+		it('shouldn’t have duplicated url', () => {
+			const s = new Set<string>()
+			ELEMENTS.forEach((artwork, index) => {
+				expect(s.has(artwork.source), artwork.css_class).to.be.false
+				s.add(artwork.source)
+			})
+		})
+		it('shouldn’t have duplicated css classes', () => {
+			const s = new Set<string>()
+			ELEMENTS.forEach((artwork, index) => {
+				expect(s.has(artwork.css_class), artwork.css_class).to.be.false
+				s.add(artwork.css_class)
+			})
+		})
 	})
 })