+++ A-apps/the-boring-rpg/client-browser/CHANGELOG.md
@@ -11,6 +11,10 @@
 - 😷 chore: some internal refactors, bump dependencies
 
 
+## v0.
+- 😍 feature: 16 new backgrounds (now totalling 114!
+- 😅 fix: login and reload features are working again
+
 ## v0.62.1
 - 👋 feature dropped: the snow for Christmas season
 - 🤩 feature: new loader